requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Authlib==1.3.0
requests==2.31.0
Werkzeug==3.0.1

# Azure SDK
azure-storage-blob==12.19.0

# Document parsing
PyMuPDF==1.23.8
python-docx==1.1.0

# Google Generative AI
google-generativeai==0.3.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# WSGI server for production
gunicorn==21.2.0